Split up projects
Closes #728
