+++ e2e/utils/registration.teardown.ts
@@ -1,0 +1,10 @@
+// Just import from playwright as we don't want to load the extension from the temp directory
+import { test as teardown } from '@playwright/test';
+
+import { cleanAuth, cleanExtension } from './loader';
+
+teardown('cleanup registration keys', async () => {
+  //   Create a new page and navigate to extension
+  await cleanAuth();
+  await cleanExtension();
+});
