+++ e2e/utils/transaction.teardown.ts
@@ -1,10 +1,9 @@
 // Just import from playwright as we don't want to load the extension from the temp directory
 import { test as teardown } from '@playwright/test';
 
-import { cleanAuth, cleanExtension } from './helper';
+import { cleanExtension } from './loader';
 
-teardown('cleanup extension storage', async () => {
+teardown('cleanup extension data', async () => {
   //   Create a new page and navigate to extension
-  await cleanAuth();
   await cleanExtension();
 });
